apps/api/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
palettum==0.3.2
gunicorn==21.2.0
python-dotenv==1.0.1
flask-cors==4.0.0
